+++ src/internal/localsession/session_logging.go
@@ -4,14 +4,16 @@ import (
 	"context"
 	"time"
 
+	"github.com/jmalloc/twelf/src/twelf"
 	"github.com/rinq/rinq-go/src/internal/attributes"
 	"github.com/rinq/rinq-go/src/rinq"
+	"github.com/rinq/rinq-go/src/rinq/constraint"
 	"github.com/rinq/rinq-go/src/rinq/ident"
 	"github.com/rinq/rinq-go/src/rinq/trace"
 )
 
 func logCreated(
-	logger rinq.Logger,
+	logger twelf.Logger,
 	id ident.SessionID,
 ) {
 	logger.Log(
@@ -21,7 +23,7 @@ func logCreated(
 }
 
 func logCall(
-	logger rinq.Logger,
+	logger twelf.Logger,
 	msgID ident.MessageID,
 	ns string,
 	cmd string,
@@ -83,7 +85,7 @@ func logCall(
 }
 
 func logAsyncRequest(
-	logger rinq.Logger,
+	logger twelf.Logger,
 	msgID ident.MessageID,
 	ns string,
 	cmd string,
@@ -107,7 +109,7 @@ func logAsyncRequest(
 
 func logAsyncResponse(
 	ctx context.Context,
-	logger rinq.Logger,
+	logger twelf.Logger,
 	msgID ident.MessageID,
 	ns string,
 	cmd string,
@@ -146,8 +148,122 @@ func logAsyncResponse(
 	}
 }
 
+func logExecute(
+	logger twelf.Logger,
+	msgID ident.MessageID,
+	ns string,
+	cmd string,
+	out *rinq.Payload,
+	err error,
+	traceID string,
+) {
+	if err != nil {
+		return // request never sent
+	}
+
+	logger.Log(
+		"%s executed '%s::%s' command (%d/o) [%s]",
+		msgID.ShortString(),
+		ns,
+		cmd,
+		out.Len(),
+		traceID,
+	)
+}
+
+func logNotify(
+	logger twelf.Logger,
+	msgID ident.MessageID,
+	ns string,
+	t string,
+	target ident.SessionID,
+	out *rinq.Payload,
+	err error,
+	traceID string,
+) {
+	if err != nil {
+		return // request never sent
+	}
+
+	logger.Log(
+		"%s sent '%s::%s' notification to %s (%d/o) [%s]",
+		msgID.ShortString(),
+		ns,
+		t,
+		target.ShortString(),
+		out.Len(),
+		traceID,
+	)
+}
+
+func logNotifyMany(
+	logger twelf.Logger,
+	msgID ident.MessageID,
+	ns string,
+	t string,
+	con constraint.Constraint,
+	out *rinq.Payload,
+	err error,
+	traceID string,
+) {
+	if err != nil {
+		return // request never sent
+	}
+
+	logger.Log(
+		"%s sent '%s::%s' notification to sessions matching %s (%d/o) [%s]",
+		msgID.ShortString(),
+		ns,
+		t,
+		con,
+		out.Len(),
+		traceID,
+	)
+}
+
+func logNotifyRecv(
+	logger twelf.Logger,
+	ref ident.Ref,
+	n rinq.Notification,
+	traceID string,
+) {
+	logger.Log(
+		"%s received '%s::%s' notification from %s (%d/i) [%s]",
+		ref.ShortString(),
+		n.Namespace,
+		n.Type,
+		n.ID.Ref.ShortString(),
+		n.Payload.Len(),
+		traceID,
+	)
+}
+
+func logListen(
+	logger twelf.Logger,
+	ref ident.Ref,
+	ns string,
+) {
+	logger.Debug(
+		"%s started listening for notifications in '%s' namespace",
+		ref.ShortString(),
+		ns,
+	)
+}
+
+func logUnlisten(
+	logger twelf.Logger,
+	ref ident.Ref,
+	ns string,
+) {
+	logger.Debug(
+		"%s stopped listening for notifications in '%s' namespace",
+		ref.ShortString(),
+		ns,
+	)
+}
+
 func logSessionDestroy(
-	logger rinq.Logger,
+	logger twelf.Logger,
 	ref ident.Ref,
 	attrs attributes.Catalog,
 	traceID string,