+++ src/internal/remotesession/client.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"sync/atomic"
 
+	"github.com/jmalloc/twelf/src/twelf"
 	opentracing "github.com/opentracing/opentracing-go"
 	"github.com/opentracing/opentracing-go/ext"
 	"github.com/rinq/rinq-go/src/internal/attributes"
@@ -11,12 +12,13 @@ import (
 	"github.com/rinq/rinq-go/src/internal/opentr"
 	"github.com/rinq/rinq-go/src/rinq"
 	"github.com/rinq/rinq-go/src/rinq/ident"
+	"github.com/rinq/rinq-go/src/rinq/trace"
 )
 
 type client struct {
 	peerID  ident.PeerID
 	invoker command.Invoker
-	logger  rinq.Logger
+	logger  twelf.Logger
 	tracer  opentracing.Tracer
 	seq     uint32
 }
@@ -24,7 +26,7 @@ type client struct {
 func newClient(
 	peerID ident.PeerID,
 	invoker command.Invoker,
-	logger rinq.Logger,
+	logger twelf.Logger,
 	tracer opentracing.Tracer,
 ) *client {
 	return &client{
@@ -45,10 +47,13 @@ func (c *client) Fetch(
 	attributes.VList,
 	error,
 ) {
+	msgID, traceID := c.nextMessageID(ctx)
+
 	span, ctx := opentr.ChildOf(ctx, c.tracer, ext.SpanKindRPCClient)
 	defer span.Finish()
 
 	opentr.SetupSessionFetch(span, ns, sessID)
+	opentr.AddTraceID(span, traceID)
 	opentr.LogSessionFetchRequest(span, keys)
 
 	out := rinq.NewPayload(fetchRequest{
@@ -58,9 +63,10 @@ func (c *client) Fetch(
 	})
 	defer out.Close()
 
-	_, in, err := c.invoker.CallUnicast(
+	in, err := c.invoker.CallUnicast(
 		ctx,
-		c.nextMessageID(),
+		msgID,
+		traceID,
 		sessID.Peer,
 		sessionNamespace,
 		fetchCommand,
@@ -97,10 +103,13 @@ func (c *client) Update(
 	attributes.VList,
 	error,
 ) {
+	msgID, traceID := c.nextMessageID(ctx)
+
 	span, ctx := opentr.ChildOf(ctx, c.tracer, ext.SpanKindRPCClient)
 	defer span.Finish()
 
 	opentr.SetupSessionUpdate(span, ns, ref.ID)
+	opentr.AddTraceID(span, traceID)
 	opentr.LogSessionUpdateRequest(span, ref.Rev, attrs)
 
 	out := rinq.NewPayload(updateRequest{
@@ -111,9 +120,10 @@ func (c *client) Update(
 	})
 	defer out.Close()
 
-	_, in, err := c.invoker.CallUnicast(
+	in, err := c.invoker.CallUnicast(
 		ctx,
-		c.nextMessageID(),
+		msgID,
+		traceID,
 		ref.ID.Peer,
 		sessionNamespace,
 		updateCommand,
@@ -161,10 +171,13 @@ func (c *client) Clear(
 	ident.Revision,
 	error,
 ) {
+	msgID, traceID := c.nextMessageID(ctx)
+
 	span, ctx := opentr.ChildOf(ctx, c.tracer, ext.SpanKindRPCClient)
 	defer span.Finish()
 
 	opentr.SetupSessionClear(span, ns, ref.ID)
+	opentr.AddTraceID(span, traceID)
 	opentr.LogSessionClearRequest(span, ref.Rev)
 
 	out := rinq.NewPayload(updateRequest{
@@ -174,9 +187,10 @@ func (c *client) Clear(
 	})
 	defer out.Close()
 
-	_, in, err := c.invoker.CallUnicast(
+	in, err := c.invoker.CallUnicast(
 		ctx,
-		c.nextMessageID(),
+		msgID,
+		traceID,
 		ref.ID.Peer,
 		sessionNamespace,
 		clearCommand,
@@ -208,10 +222,13 @@ func (c *client) Destroy(
 	ctx context.Context,
 	ref ident.Ref,
 ) error {
+	msgID, traceID := c.nextMessageID(ctx)
+
 	span, ctx := opentr.ChildOf(ctx, c.tracer, ext.SpanKindRPCClient)
 	defer span.Finish()
 
 	opentr.SetupSessionDestroy(span, ref.ID)
+	opentr.AddTraceID(span, traceID)
 	opentr.LogSessionDestroyRequest(span, ref.Rev)
 
 	out := rinq.NewPayload(destroyRequest{
@@ -220,9 +237,10 @@ func (c *client) Destroy(
 	})
 	defer out.Close()
 
-	_, in, err := c.invoker.CallUnicast(
+	in, err := c.invoker.CallUnicast(
 		ctx,
-		c.nextMessageID(),
+		msgID,
+		traceID,
 		ref.ID.Peer,
 		sessionNamespace,
 		destroyCommand,
@@ -241,8 +259,14 @@ func (c *client) Destroy(
 	return nil
 }
 
-func (c *client) nextMessageID() ident.MessageID {
+func (c *client) nextMessageID(ctx context.Context) (msgID ident.MessageID, traceID string) {
 	seq := atomic.AddUint32(&c.seq, 1)
+	msgID = c.peerID.Session(0).At(0).Message(seq)
+	traceID = trace.Get(ctx)
 
-	return c.peerID.Session(0).At(0).Message(seq)
+	if traceID == "" {
+		traceID = msgID.String()
+	}
+
+	return
 }