+++ src/internal/remotesession/server.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"errors"
 
+	"github.com/jmalloc/twelf/src/twelf"
 	opentracing "github.com/opentracing/opentracing-go"
 	"github.com/rinq/rinq-go/src/internal/attributes"
 	"github.com/rinq/rinq-go/src/internal/command"
@@ -11,12 +12,13 @@ import (
 	"github.com/rinq/rinq-go/src/internal/opentr"
 	"github.com/rinq/rinq-go/src/rinq"
 	"github.com/rinq/rinq-go/src/rinq/ident"
+	"github.com/rinq/rinq-go/src/rinq/trace"
 )
 
 type server struct {
 	peerID   ident.PeerID
 	sessions *localsession.Store
-	logger   rinq.Logger
+	logger   twelf.Logger
 }
 
 // Listen attaches a new remote session service to the given command server.
@@ -24,7 +26,7 @@ func Listen(
 	svr command.Server,
 	peerID ident.PeerID,
 	sessions *localsession.Store,
-	logger rinq.Logger,
+	logger twelf.Logger,
 ) error {
 	s := &server{
 		peerID:   peerID,
@@ -75,6 +77,7 @@ func (s *server) fetch(
 	sessID := s.peerID.Session(args.Seq)
 
 	opentr.SetupSessionFetch(span, args.Namespace, sessID)
+	opentr.AddTraceID(span, trace.Get(ctx))
 	opentr.LogSessionFetchRequest(span, args.Keys)
 
 	sess, ok := s.sessions.Get(sessID)
@@ -123,6 +126,7 @@ func (s *server) update(
 	sessID := s.peerID.Session(args.Seq)
 
 	opentr.SetupSessionUpdate(span, args.Namespace, sessID)
+	opentr.AddTraceID(span, trace.Get(ctx))
 	opentr.LogSessionUpdateRequest(span, args.Rev, args.Attrs)
 
 	sess, ok := s.sessions.Get(sessID)
@@ -180,6 +184,7 @@ func (s *server) clear(
 	sessID := s.peerID.Session(args.Seq)
 
 	opentr.SetupSessionClear(span, args.Namespace, sessID)
+	opentr.AddTraceID(span, trace.Get(ctx))
 	opentr.LogSessionClearRequest(span, args.Rev)
 
 	sess, ok := s.sessions.Get(sessID)
@@ -228,6 +233,7 @@ func (s *server) destroy(
 	sessID := s.peerID.Session(args.Seq)
 
 	opentr.SetupSessionDestroy(span, sessID)
+	opentr.AddTraceID(span, trace.Get(ctx))
 	opentr.LogSessionDestroyRequest(span, args.Rev)
 
 	sess, ok := s.sessions.Get(sessID)