+++ src/rinqamqp/internal/amqputil/trace.go
@@ -7,23 +7,16 @@ import (
 	"github.com/streadway/amqp"
 )
 
-// PackTrace sets msg.CorrelationId to the trace ID in ctx, and returns the ID.
-//
-// If ctx does not have a trace ID, the return value is msg.MessageId.
+// PackTrace sets msg.CorrelationId to traceID, only if it differs to msgID.
 //
 // The AMQP correlation ID field is used to tie "root" requests (be they command
 // requests or notifications) to any requests that are made in response to that
 // "root" request. This is different to the popular use of the correlation ID
 // field, which is often used to relate a response to a request.
-func PackTrace(ctx context.Context, msg *amqp.Publishing) string {
-	traceID := trace.Get(ctx)
-
-	if traceID == "" || traceID == msg.MessageId {
-		return msg.MessageId
+func PackTrace(msg *amqp.Publishing, traceID string) {
+	if traceID != msg.MessageId {
+		msg.CorrelationId = traceID
 	}
-
-	msg.CorrelationId = traceID
-	return traceID
 }
 
 // UnpackTrace creates a new context with a trace ID based on the AMQP correlation