+++ src/rinqamqp/internal/notifyamqp/notifier.go
@@ -3,6 +3,7 @@ package notifyamqp
 import (
 	"context"
 
+	"github.com/jmalloc/twelf/src/twelf"
 	"github.com/rinq/rinq-go/src/internal/notify"
 	"github.com/rinq/rinq-go/src/internal/service"
 	"github.com/rinq/rinq-go/src/rinq"
@@ -18,14 +19,14 @@ type notifier struct {
 
 	peerID   ident.PeerID
 	channels amqputil.ChannelPool
-	logger   rinq.Logger
+	logger   twelf.Logger
 }
 
 // newNotifier creates, initializes and returns a new notifier.
 func newNotifier(
 	peerID ident.PeerID,
 	channels amqputil.ChannelPool,
-	logger rinq.Logger,
+	logger twelf.Logger,
 ) notify.Notifier {
 	n := &notifier{
 		peerID:   peerID,
@@ -44,18 +45,18 @@ func newNotifier(
 func (n *notifier) NotifyUnicast(
 	ctx context.Context,
 	msgID ident.MessageID,
+	traceID string,
 	target ident.SessionID,
 	ns string,
 	notificationType string,
 	payload *rinq.Payload,
-) (traceID string, err error) {
+) (err error) {
 	msg := amqp.Publishing{
 		MessageId: msgID.String(),
 	}
 
-	packCommonAttributes(&msg, ns, notificationType, payload)
+	packCommonAttributes(&msg, traceID, ns, notificationType, payload)
 	packTarget(&msg, target)
-	traceID = amqputil.PackTrace(ctx, &msg)
 
 	err = amqputil.PackSpanContext(ctx, &msg)
 
@@ -69,18 +70,18 @@ func (n *notifier) NotifyUnicast(
 func (n *notifier) NotifyMulticast(
 	ctx context.Context,
 	msgID ident.MessageID,
+	traceID string,
 	con constraint.Constraint,
 	ns string,
 	notificationType string,
 	payload *rinq.Payload,
-) (traceID string, err error) {
+) (err error) {
 	msg := amqp.Publishing{
 		MessageId: msgID.String(),
 	}
 
-	packCommonAttributes(&msg, ns, notificationType, payload)
+	packCommonAttributes(&msg, traceID, ns, notificationType, payload)
 	packConstraint(&msg, con)
-	traceID = amqputil.PackTrace(ctx, &msg)
 
 	err = amqputil.PackSpanContext(ctx, &msg)