+++ src/rinqamqp/peer.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"sync/atomic"
 
+	"github.com/jmalloc/twelf/src/twelf"
 	opentracing "github.com/opentracing/opentracing-go"
 	"github.com/rinq/rinq-go/src/internal/command"
 	"github.com/rinq/rinq-go/src/internal/localsession"
@@ -31,7 +32,7 @@ type peer struct {
 	server      command.Server
 	notifier    notify.Notifier
 	listener    notify.Listener
-	logger      rinq.Logger
+	logger      twelf.Logger
 	tracer      opentracing.Tracer
 
 	seq        uint32
@@ -47,7 +48,7 @@ func newPeer(
 	server command.Server,
 	notifier notify.Notifier,
 	listener notify.Listener,
-	logger rinq.Logger,
+	logger twelf.Logger,
 	tracer opentracing.Tracer,
 ) *peer {
 	p := &peer{
@@ -114,12 +115,15 @@ func (p *peer) Listen(ns string, handler rinq.CommandHandler) error {
 		) {
 			span := opentracing.SpanFromContext(ctx)
 
+			traceID := trace.Get(ctx)
+
 			opentr.SetupCommand(
 				span,
 				req.ID,
 				req.Namespace,
 				req.Command,
 			)
+			opentr.AddTraceID(span, traceID)
 			opentr.LogServerRequest(span, p.id, req.Payload)
 
 			handler(
@@ -129,7 +133,7 @@ func (p *peer) Listen(ns string, handler rinq.CommandHandler) error {
 					req,
 					res,
 					p.id,
-					trace.Get(ctx),
+					traceID,
 					p.logger,
 					span,
 				),