+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ## Next Release
 
+- **[IMPROVED]** `Revision.Refresh()` always returns a usable revision (outside of a network error)
+
+## 0.7.0 (2018-02-03)
+
 - **[BC]** `Session.CurrentRevision()` no longer returns an error
 - **[BC]** `Session.Destroy()` no longer waits for pending calls, use `Session.Done()` to wait
 - **[BC]** Remove `Revision.Ref()` method
+- **[FIX]** Constraint serialisation no longer returns an in-use buffer to the pool
 - **[NEW]** Add `Revision.SessionID()` method
 
 ## 0.6.0 (2017-10-27)